Autclose for streams
